--- Source.cpp
@@ -1,202 +1,0 @@
-#include <iostream>
-#include <vector>
-#include <algorithm>
-using namespace std;
-
-
-template<class T>
-class HeapOverArray {
-    std::vector<T> v;
-
-    //скрытый интерфейс
-    void HeapMaxify(size_t ind, size_t size = v.size())
-    {
-
-        //https://habr.com/ru/post/587228/
-
-        //определяем левый и правый дочернии узлы
-        size_t  left   = 2 * ind
-                ,right = left + 1;
-
-        //максимальное, как мы думаем, значение
-        size_t largest = ind;
-
-        //если левый дочерний узел больше чем корень
-        //поддерева(корень с индексом largest)
-        //то мы изменяем largest на индекс левого
-        //или правого дочернего узла
-        if (left < size && v[left] > v[largest]) {
-            largest = left;
-        }
-        if (right < size && v[right] > v[largest]) {
-            largest = right;
-        }
-
-        //если всё-таки значение изначально проверяемого 
-        //узла с индексом i не является самым большим
-        //то проверяемый узел и узел с большим значением меняются местами
-        if (largest != ind) {
-            std::swap(v, ind, largest);
-
-            //Рекурсивно вызываем функцию, чтобы в случае чего
-            //отправить элемент ниже по пирамиде
-            HeapMaxify(largest, size);
-        }
-    }
-
-public:
-    HeapOverArray() {}
-    HeapOverArray(const std::vector<T>& initv) : v(initv) {}    /// Требуется, чтобы массив был заранее упорядочен
-
-    const std::vector<T>& getVector() const { return v; }
-
-    /// Если такой элемент уже существует, то изменения не вносятся, а функция возвращает false.
-    /// Если элемент добавлен, то функция возвращает true.
-       
-
-        bool addNode(const T & o) {
-            /// 1. Определяем место вставки (первое свободное место в нижнем горизонтальном ряду).
-       /// 2. Сравниваем с родителем, если значение больше родительского, меняем местами с родителем (в цикле).
-            size_t pos = v.size();
-
-            if (pos == 0)
-            {
-                v.push_back(o);
-                return true;
-            }
-
-            size_t parent_pos = (pos + 1) / 2 - 1;
-
-            while (parent_pos >= 0 && pos > 0) {
-                if (v[pos] > v[parent_pos])
-                {
-                    swap(v.at(pos), v.at(parent_pos));
-                }
-                pos = parent_pos;
-                parent_pos = (pos + 1) / 2 - 1;
-            }
-        }
-
-    /// ЗАДАЧИ, ч.2
-    ///  1. Удаление элемента из двоичной кучи
-    ///  2. Тестирование корректности реализации
-    ///  3. Построение итератора для двоичной кучи
-    ///
-
-    bool removeNode(const T& o) 
-    {
-        auto it = std::find(v.begin(), v.end(), o); //возвращаем итератор на нужный элемент
-
-        if (it != v.end()) //удаляем не последний элемент
-        {
-            //просто меняем наш элемент с последним местами и удаляем
-            std::swap(it, v.rbegin());
-            v.pop_back();
-            v.shrink_to_fit(); //обрезали массив
-
-            //восстанавливаем порядок ???
-
-            std::sort_heap(v.begin(), v.end())
-
-            
-        }
-            v.pop_back(); 
-            v.shrink_to_fit(); //очищаем неиспользованное пространство
-    }
-
-    class iterator {
-        std::vector<T>* v;
-        size_t idx;
-
-    public:
-        iterator() : v(nullptr) {}
-        iterator(std::vector<T>* v_, size_t idx_) : v(v_), idx(idx_) {}
-
-        ///
-        /// 100
-        ///  70          60
-        ///  25    10    40     5
-        ///   8  2  9  1  3
-        ///
-        /// 100 70 60 25 10 40 5 8 2 9 1 3
-        ///
-
-        iterator& operator++() {
-            /// 1. Узнаём высоту элемента
-            /// 2. Узнаём, слева мы или справа
-
-            /// Если двигаемся вверх (если нет дочернего элемента справа),
-            /// делим номер в дереве пополам,
-            /// затем отбрасывая дробную часть, в цикле пока не дойдём до чётного элемента,
-            /// затем ещё раз делим номер пополам.
-            ///
-            /// Если двигаемся вниз, то: 1) увеличиваем номер в 2 раза и прибавляем 1;
-            /// 2) пока есть дочерние элементы, увеличиваем номер в 2 раза.
-            ///
-            /// Если двинуться некуда, превращаем итератор в невалидный путём зануления v.
-            ///
-            /// Лист == номер элемента, умноженный в 2 раза, меньше количества элементов в дереве
-
-            return *this;
-        }
-
-        iterator operator++(int) {
-            auto prev = *this;
-            ++(*this);
-            return prev;
-        }
-
-        operator bool() const { return v != nullptr; }
-        /// Теперь можно писать так:
-        ///
-        /// HeapOverArray::iterator iter;
-        /// if (iter) { /* ... */ }
-        ///
-        /// HeapOverArray heap;
-        /// for (auto it = heap.begin(); it; it++) { /* ... */ }
-
-        const T& operator*() const { return v[idx]; }
-        T& operator*() { return v[idx]; }
-    };
-
-    iterator begin() {
-        if (v.empty())
-            return iterator();
-        /// От корня двигаемся влево до упора
-        /// TODO
-    }
-
-    iterator end() { return iterator(); }
-
-    ///   #1 #2 #3 #4
-    ///    7  3  4
-    ///    7  3  4  8
-    ///    7  8  4  3
-    ///    8  7  4  3
-    ///    8  4  7  3
-};
-
-template<class T>
-bool testHeapAdd(const std::vector<T>& initial, const T& value, const std::vector<T>& expected) {
-    HeapOverArray<T> heap(initial);
-    heap.addNode(value);
-    auto v = heap.getVector();
-    if (v.size() != expected.size()) {
-        std::cerr << "size difference: expected " << expected.size() << ", got " << v.size() << std::endl;
-        return false;
-    }
-    for (size_t i = 0; i < expected.size(); i++)
-        if (v[i] != expected[i]) {
-            std::cerr << "difference in " << i << ": expected " << expected[i] << ", got " << v[i] << std::endl;
-            return false;
-        }
-    return true;
-}
-
-int main()
-{
-    vector<int> initial = { 7, 3, 4 };
-    vector<int> expected = { 8, 4, 7, 3 };
-    testHeapAdd(initial, 8, expected);
-    return 0;
-}
